requirements.txt
dendropy>=4.0.0
scipy>=1.0.0
numpy>=1.11.0
matplotlib
rtree
gdal
specify-lmpy>=3.1.19
